+++ silzila-frontend/src/Components/DataConnection/DataConnection.tsx
@@ -3,53 +3,26 @@
 // Creating new and editing existing connections are handled in FormDialog child component
 
 import React, { useEffect, useState } from "react";
-import { connect } from "react-redux";
 import { VisibilitySharp } from "@mui/icons-material";
 import { Tooltip } from "@mui/material";
 import { SelectListItem } from "../CommonFunctions/SelectListItem";
-import { Dispatch } from "redux";
-import { NotificationDialog } from "../CommonFunctions/DialogComponents";
-import FormDialog from "./FormDialog";
-import { DataConnectionDetails, DataConnectionProps } from "./DataConnectionInterfaces";
 import { isLoggedProps } from "../../redux/UserInfo/IsLoggedInterfaces";
 import FetchData from "../ServerCall/FetchData";
-import { setDataConnectionListToState } from "../../redux/DataSet/datasetActions";
 import { ConnectionItem } from "../../redux/DataSet/DatasetStateInterfaces";
-import { resetAllStates } from "../../redux/TabTile/TabTileActionsAndMultipleDispatches";
 import AddIcon from "@mui/icons-material/Add";
 import StorageOutlinedIcon from "@mui/icons-material/StorageOutlined";
-import { AlertColor } from "@mui/material/Alert";
 import Logger from "../../Logger";
-
-const initialState = {
-	vendor: "",
-	vendorError: "",
-	server: "",
-	serverError: "",
-	port: "",
-	portError: "",
-	database: "",
-	databaseError: "",
-	username: "",
-	userNameError: "",
-	connectionName: "",
-	connectionNameError: "",
-	password: "",
-	passwordError: "",
-	httpPath: "",
-	httpPathError:"",
-};
+import { useNavigate } from "react-router-dom";
+import { connect } from "react-redux";
+import { Dispatch } from "redux";
+import { DataConnectionProps } from "./DataConnectionInterfaces";
+import { setDataConnectionListToState } from "../../redux/DataSet/datasetActions";
+import { resetAllStates } from "../../redux/TabTile/TabTileActionsAndMultipleDispatches";
 
 const DataConnection = (props: DataConnectionProps) => {
 	const [dataConnectionList, setDataConnectionList] = useState<ConnectionItem[]>([]);
-	const [showForm, setShowForm] = useState<boolean>(false);
-	const [regOrUpdate, setRegOrUpdate] = useState<string>("Register");
-	const [account, setAccount] = useState<DataConnectionDetails>(initialState);
-	const [dataConnId, setDataConnId] = useState<string>("");
-	const [viewMode, setViewMode] = useState<boolean>(false);
-	const [severity, setSeverity] = useState<AlertColor>("success");
-	const [openAlert, setOpenAlert] = useState<boolean>(false);
-	const [testMessage, setTestMessage] = useState<string>("Testing alert");
+	const [mode, setMode] = useState<string>("New");
+    const navigate = useNavigate();
 
 	useEffect(() => {
 		props.resetAllStates();
@@ -57,7 +30,8 @@ const DataConnection = (props: DataConnectionProps) => {
 		// eslint-disable-next-line
 	}, []);
 
-	// Get Info on DataConnection from server
+    
+  // Get Info on DataConnection from server
 	const getInformation = async () => {
 		var result: any = await FetchData({
 			requestType: "noData",
@@ -75,178 +49,6 @@ const DataConnection = (props: DataConnectionProps) => {
 		}
 	};
 
-	// ================================= when newButton clicked ====================
-
-	//=============== set Mode ===============================
-	// TODO:need to specify types
-	const handleMode = (mode: string) => {
-		if (mode === "New") {
-			setRegOrUpdate("Register");
-		} else if (mode === "Edit") {
-			setAccount({ ...account, password: "" });
-			setRegOrUpdate("Update");
-		}
-	};
-
-	// =======================================
-	// open Form
-	//  ======================================
-
-	const showAndHideForm = () => {
-		if (showForm === true) {
-			setShowForm(false);
-			setAccount(initialState);
-			setDataConnId("");
-			setViewMode(false);
-		} else {
-			setShowForm(true);
-		}
-	};
-
-	// ========================================================================
-
-	// ==================================================
-	// when Visibility icon Clicked
-	// ==================================================
-	const ViewOrEditDc = async (dcuid: string) => {
-		setDataConnId(dcuid);
-		// TODO need to specify type
-		var result: any = await FetchData({
-			requestType: "noData",
-			method: "GET",
-			url: "database-connection/" + dcuid,
-			headers: { Authorization: `Bearer ${props.token}` },
-		});
-
-		if (result.status) {
-			setAccount({ ...result.data, password: "*******" });
-			setShowForm(true);
-			setViewMode(true);
-		} else {
-			Logger("error", result.data.detail);
-		}
-	};
-
-	// ==============================================================
-	//  Register dc
-	//  ==============================================================
-
-	const handleRegister = async () => {
-		var data:any = {
-			vendor: account.vendor,
-			server: account.server,
-			port: account.port,
-			database: account.database,
-			password: account.password,
-			connectionName: account.connectionName,
-		};
-		if(account.vendor === "databricks"){
-           data.httpPath = account.httpPath;
-		}else{
-			data.username = account.username;
-		}
-		// TODO need to specify type
-		var response: any = await FetchData({
-			requestType: "withData",
-			method: "POST",
-			url: "database-connection",
-			headers: { "Content-Type": "application/json", Authorization: `Bearer ${props.token}` },
-			data: data,
-		});
-
-		if (response.status) {
-			if (response.data.message === "Friendlly Name is already used") {
-				setAccount({
-					...account,
-					connectionNameError: "Friendlly Name is already used try any other Name",
-				});
-			} else {
-				setOpenAlert(true);
-				setSeverity("success");
-				setTestMessage("Data Connection successful");
-				getInformation();
-				setTimeout(() => {
-					setOpenAlert(false);
-					setTestMessage("");
-					setShowForm(false);
-					setAccount(initialState);
-				}, 3000);
-			}
-		} else {
-			Logger("error", response);
-		}
-	};
-
-	// ==============================================================
-	// Update Dc
-	// ==============================================================
-	const handleonUpdate = async () => {
-		var data:any = {
-			vendor: account.vendor,
-			server: account.server,
-			port: account.port,
-			database: account.database,
-			password: account.password,
-			connectionName: account.connectionName,
-		};
-		if(account.vendor === "databricks"){
-			data.httpPath = account.httpPath;
-		 }else{
-			 data.username = account.username;
-		 }
-		// TODO need to specify type
-		var response: any = await FetchData({
-			requestType: "withData",
-			method: "PUT",
-			url: "database-connection/" + dataConnId,
-			headers: { "Content-Type": "application/json", Authorization: `Bearer ${props.token}` },
-			data: data,
-		});
-
-		if (response.status) {
-			setSeverity("success");
-			setOpenAlert(true);
-			setTestMessage("Updated Successfully!");
-			setTimeout(() => {
-				setOpenAlert(false);
-				setTestMessage("");
-				showAndHideForm();
-				getInformation();
-			}, 3000);
-		} else {
-			setSeverity("error");
-			setOpenAlert(true);
-			setTestMessage(response.data.detail);
-			// setTimeout(() => {
-			// 	setOpenAlert(false);
-			// 	setTestMessage("");
-			// }, 3000);
-		}
-	};
-
-	// ===========================================================
-	// props to form Component
-	// ===========================================================
-
-	const properties = {
-		account,
-		setAccount,
-		viewMode,
-		setViewMode,
-		showForm,
-		showAndHideForm,
-		regOrUpdate,
-		handleMode,
-		token: props.token,
-		setSeverity,
-		setOpenAlert,
-		setTestMessage,
-		dataConnId,
-		handleRegister,
-		getInformation,
-		handleonUpdate,
-	};
-
 	return (
 		<div className="dataConnectionContainer">
 			<div className="containersHead">
@@ -258,9 +60,8 @@ const DataConnection = (props: DataConnectionProps) => {
 					className="containerButton"
 					onClick={(e: any) => {
 						Logger("info", "add new connection");
-						handleMode("New");
-						showAndHideForm();
-					}}
+						navigate("/newdataconnection", {state: { mode: mode}});
+						}}
 					title="Create New DB Connection"
 				>
 					<AddIcon />
@@ -298,7 +99,8 @@ const DataConnection = (props: DataConnectionProps) => {
 														height: "1rem",
 														margin: "auto 7px auto auto",
 													}}
-													onClick={() => ViewOrEditDc(dc.id)}
+													onClick={() => 
+														navigate("/newdataconnection", {state: { id: dc.id, value: dc.vendor}})}
 												/>
 											</Tooltip>
 										) : null}
@@ -308,33 +110,22 @@ const DataConnection = (props: DataConnectionProps) => {
 						);
 					})}
 			</div>
-			<FormDialog {...properties} />
-
-			{/* Alert to display success / failure info */}
-			<NotificationDialog
-				onCloseAlert={() => {
-					setOpenAlert(false);
-					setTestMessage("");
-				}}
-				severity={severity}
-				testMessage={testMessage}
-				openAlert={openAlert}
-			/>
 		</div>
 	);
 };
 
 const mapStateToProps = (state: isLoggedProps) => {
-	return {
-		token: state.isLogged.accessToken,
+return {
+	token: state.isLogged.accessToken,
 	};
-};
+	};
 
-const mapDispatchToProps = (dispatch: Dispatch<any>) => {
-	return {
+	const mapDispatchToProps = (dispatch: Dispatch<any>) => {
+		return {
 		resetAllStates: () => dispatch(resetAllStates()),
 		setDataConnectionListToState: (list: ConnectionItem[]) =>
-			dispatch(setDataConnectionListToState(list)),
-	};
-};
+		dispatch(setDataConnectionListToState(list)),
+		};
+		};
+
 export default connect(mapStateToProps, mapDispatchToProps)(DataConnection);
